Python Project
This online recruitment system was done using python language. There was 3 module - profile,
vacancy, and status. From the profile module, the applicant can make their profile and modify the
profile. In a vacancy, there is all the available circular and from that, the applicant can apply for the
ob. In the status module, the user can see if their application is accepted or not, if it is accepted then
the applicant will get the written exam link and also the viva link